Don't pollute environment with undefined BABEL_ENV
Since the default value for setBabelEnv is `true`, the code attempts to set `process.env.BABEL_ENV` for transpiling the package, and restore it to the previous value when done. The problem is that when it wasn't defined in the first place it gets set to `undefined`, as in `process.env.BABEL_ENV === 'undefined'`. The causes a multitude of bizarre issues as you can imagine.

Fixes https://github.com/AtomLinter/linter-eslint/issues/921.